+++ breeze-reflect/src/main/kotlin/com/windea/breezeframework/reflect/extensions/java/ExecutableExtensions.kt
@@ -1,6 +1,6 @@
 @file:Reference("[klutter](https://github.com/kohesive/klutter/blob/master/reflect/src/main/kotlin/uy/klutter/reflect/Helpers.kt)")
 
-package com.windea.breezeframework.reflect.extensions
+package com.windea.breezeframework.reflect.extensions.java
 
 import com.windea.breezeframework.core.annotations.internal.*
 import java.lang.reflect.*